requirements.txt
numpy==2.4.3
Pillow==12.1.1
PyQt6==6.10.2
pyqt6_sip==13.11.0
pyqtgraph==0.14.0
